Got rid of some warnings and implemented jacobian stuff
git-svn-id: https://dune.mathematik.uni-freiburg.de/svn/dune-fem/trunk@86 4a6c9bf5-c326-0410-8cc6-c7cabc8ddcf0

+++ fem/pass/selection.hh
@@ -68,8 +68,10 @@ namespace Dune {
 
   template <class Head, class Tail>
   struct MaxIndex<Pair<Head, Tail> > {
-    enum { value = (Head::value > MaxIndex<Tail>::value ?
-                    Head::value : MaxIndex<Tail>::value) };
+    enum { value = (static_cast<int>(Head::value) > 
+                    static_cast<int>(MaxIndex<Tail>::value) ?
+                    static_cast<int>(Head::value) : 
+                    static_cast<int>(MaxIndex<Tail>::value)) };
   };
 
   template <>
